+++ internal/update/replace_unix.go
@@ -1,0 +1,9 @@
+//go:build !windows
+
+package update
+
+import "os"
+
+func replaceExecutable(staged, target string) error {
+	return os.Rename(staged, target)
+}